refactor, add formulas for area calculation

+++ src/main/java/edu/shapes/interfaces/Shape.java
@@ -2,5 +2,5 @@ package edu.shapes.interfaces;
 
 public interface Shape {
 	
-	public float getArea(); 
+	public double getArea(); 
 }
